refactor(agent): extract runPassThrough handler (PR-2a slice 3)
